Added LineDirection and Disturbance import service

+++ src/main/java/be/vb/storingenmelder/repository/CityRepository.java
@@ -3,5 +3,5 @@ package be.vb.storingenmelder.repository;
 import be.vb.storingenmelder.domain.City;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CityRepository extends JpaRepository<City, Integer> {
+public interface CityRepository extends JpaRepository<City, Long> {
 }
