+++ src/main/java/be/vb/storingenmelder/repository/LineRepository.java
@@ -3,5 +3,5 @@ package be.vb.storingenmelder.repository;
 import be.vb.storingenmelder.domain.Line;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface LineRepository extends JpaRepository<Line, Integer> {
+public interface LineRepository extends JpaRepository<Line, Long> {
 }
